[SRK-22] Add tests to remaining modules
This commit will add tests to the modules that were missing those,
when it made sense to have them.

It also completes the 'Variant' arbitrary instance, but identity tests
for this type are commented out because of the intricacies of the
'MessagePack' and 'To/FromJSON' instances. The 'Binary' instance does
not work either, for reasons that require further investigation.

A bug has also been fixed in 'Serokell.Util.Common', in
'indexedSubList', and the function«s description has been updated.